+++ agr_literature_service/lit_processing/data_ingest/dqm_ingest/sort_dqm_json_reference_updates.py
@@ -13,7 +13,7 @@ from agr_literature_service.api.models import ReferenceModel, ModReferenceTypeMo
     ModCorpusAssociationModel, AuthorModel, CrossReferenceModel, ModModel
 from agr_literature_service.lit_processing.data_ingest.dqm_ingest.utils.dqm_processing_utils import \
     compare_authors_or_editors
-from agr_literature_service.lit_processing.utils.file_processing_utils import split_identifier, write_json
+from agr_literature_service.lit_processing.utils.generic_utils import split_identifier, write_json
 from agr_literature_service.lit_processing.utils.sqlalchemy_utils import create_postgres_session,\
     create_postgres_engine, sqlalchemy_load_ref_xref
 from agr_literature_service.lit_processing.utils.email_utils import send_email
