+++ agr_literature_service/lit_processing/data_ingest/pubmed_ingest/sanitize_pubmed_json.py
@@ -1,7 +1,7 @@
 import json
 from os import environ, makedirs, path
 
-from agr_literature_service.lit_processing.utils.file_processing_utils import write_json
+from agr_literature_service.lit_processing.utils.generic_utils import write_json
 from agr_literature_service.lit_processing.utils.tmp_files_utils import init_tmp_dir
 
 init_tmp_dir()
