+++ agr_literature_service/lit_processing/tests/functional_tests.py
@@ -16,7 +16,7 @@ from agr_literature_service.lit_processing.tests.mod_populate_load import post_m
 from agr_literature_service.lit_processing.tests.parse_pubmed_json_reference import parse_pubmed_json_reference
 from agr_literature_service.lit_processing.tests.process_many_pmids_to_json import process_many_pmids_to_json
 from agr_literature_service.lit_processing.utils.sqlalchemy_utils import sqlalchemy_load_ref_xref
-from agr_literature_service.lit_processing.utils.file_processing_utils import split_identifier
+from agr_literature_service.lit_processing.utils.generic_utils import split_identifier
 
 from agr_literature_service.lit_processing.tests.generate_dqm_json_test_set import load_sample_json
 from agr_literature_service.lit_processing.data_ingest.pubmed_ingest.xml_to_json import generate_json
